feat: l1_fetcher scaffolding
