+++ state-reconstruct-fetcher/src/constants.rs
@@ -8,6 +8,10 @@ pub mod ethereum {
     /// Block number in Ethereum of the first Boojum-formatted block.
     pub const BOOJUM_BLOCK: u64 = 18_715_403;
 
+    /// Block number in Ethereum of the first block storing pubdata within blobs.
+    // NOTE: Not actual number, TBD.
+    pub const BLOB_BLOCK: u64 = 20_715_403;
+
     /// zkSync smart contract address.
     pub const ZK_SYNC_ADDR: &str = "0x32400084C286CF3E17e7B677ea9583e60a000324";
 }
