+++ state-reconstruct-fetcher/src/types/common.rs
@@ -1,10 +1,90 @@
 //! A collection of functions that get reused throughout format versions.
-// TODO: Move these to a dedicated parser struct.
-use ethers::types::U256;
+use ethers::{abi, types::U256};
 
 use super::{PackingType, ParseError};
 use crate::constants::zksync::{LENGTH_BITS_OFFSET, OPERATION_BITMASK};
 
+pub struct ExtractedToken {
+    pub new_l2_block_number: U256,
+    pub timestamp: U256,
+    pub new_enumeration_index: U256,
+    pub state_root: Vec<u8>,
+    pub number_of_l1_txs: U256,
+    pub priority_operations_hash: Vec<u8>,
+    pub system_logs: Vec<u8>,
+    pub total_l2_to_l1_pubdata: Vec<u8>,
+}
+
+// TODO: Maybe consider using an enum for post-boojum formats.
+// I.e. V2::OnlyCalldata & V2::CalldataAndBlob.
+impl TryFrom<&abi::Token> for ExtractedToken {
+    type Error = ParseError;
+
+    fn try_from(token: &abi::Token) -> Result<Self, Self::Error> {
+        let abi::Token::Tuple(block_elems) = token else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "struct elements".to_string(),
+            ));
+        };
+
+        let abi::Token::Uint(new_l2_block_number) = block_elems[0].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "blockNumber".to_string(),
+            ));
+        };
+
+        let abi::Token::Uint(timestamp) = block_elems[1].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo("timestamp".to_string()));
+        };
+
+        let abi::Token::Uint(new_enumeration_index) = block_elems[2].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "indexRepeatedStorageChanges".to_string(),
+            ));
+        };
+
+        let abi::Token::FixedBytes(state_root) = block_elems[3].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "newStateRoot".to_string(),
+            ));
+        };
+
+        let abi::Token::Uint(number_of_l1_txs) = block_elems[4].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "numberOfLayer1Txs".to_string(),
+            ));
+        };
+
+        let abi::Token::FixedBytes(priority_operations_hash) = block_elems[5].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "priorityOperationsHash".to_string(),
+            ));
+        };
+
+        let abi::Token::Bytes(system_logs) = block_elems[8].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo("systemLogs".to_string()));
+        };
+
+        let abi::Token::Bytes(total_l2_to_l1_pubdata) = block_elems[9].clone() else {
+            return Err(ParseError::InvalidCommitBlockInfo(
+                "totalL2ToL1Pubdata".to_string(),
+            ));
+        };
+
+        Ok(Self {
+            new_l2_block_number,
+            timestamp,
+            new_enumeration_index,
+            state_root,
+            number_of_l1_txs,
+            priority_operations_hash,
+            system_logs,
+            total_l2_to_l1_pubdata,
+        })
+    }
+}
+
+// TODO: Move these to a dedicated parser struct.
 pub fn read_compressed_value(bytes: &[u8], pointer: &mut usize) -> Result<PackingType, ParseError> {
     let metadata = u8::from_be_bytes(read_next_n_bytes(bytes, pointer));
     let operation = metadata & OPERATION_BITMASK;
